// src/app/Data/Interfaces/IMovieOrShow.ts
export  interface IMovieOrShow {    
       movieOrShowId: number;
       movieOrShowName: string;
       movieOrShowTitle: string;
       movieOrShowDescription: string;
       ratingIMDB: number;
       language: string;
       genre: string;
       isMovie: boolean;
       imagePath: string;  
       isPrimeShow: boolean;  
    }